Strip invalid characters from the generated Nonce.
By introducing the Base64 encoding, we inadvertently introduced the potential for `+` `/` and `=` characters that weren't present in the `Guid.NewGuid("N")` implementation. The `+` specifically wasn't round-tripping from the identity provider, and it randomly causes an `Invalid nonce` message when attempting to login depending on the characters present in the nonce.

Fixes OctopusDeploy/Issues#2884