Now handling all error cases as BadRequests.
I think this is semantically correct. As it turns out, I also think using the query string as a way to pass information back from the server to the HTML/JS application has too many negative side-effects. For example: if a deep-link has a query string, or if the user attempts multiple times to sign in you end up with duplicates.